Added constructor method for vector slice types
'new()' method for VectorSlice and MutVectorSlice need to do
check with the given range for the slice. For this, this method
must return a result.

The Result-less alternative forces the user to create a vector
slice with the range being the entire original range, letting the
slice be a mutable variable, and then using the 'slice_range'
method to change it to the desired slice.

The 'slice_range' method must return a Result anyway, and anyone
who wants to create a slice of a vector with the entire range of
the given vector can just unwrap the Result safely.