Changed crate name from adv_linalg to adv_linalg_core
This is in preparation for the multiple crates that is
going to be used for this project.

The reason for the multiple crates is that procedral macros
are store as an independent crate. However, the main library
will have procedral macros like attributes, but will operate on
objects from the core. So this forces the use of multiple crates.